read_word/requirements.txt
bcrypt==4.1.2
cffi==1.16.0
cryptography==41.0.7
dnspython==2.4.2
paramiko==3.4.0
pycparser==2.21
pymongo==4.6.1
PyNaCl==1.5.0
python-dotenv==1.0.0
schedule==1.2.1
sshtunnel==0.4.0
